+++ src/vec4/vec4_len.c
@@ -6,7 +6,7 @@
 /*   By: rshin <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/02 15:49:12 by rshin             #+#    #+#             */
-/*   Updated: 2025/10/07 19:02:47 by rshin            ###   ########.fr       */
+/*   Updated: 2025/10/14 20:10:22 by rshin            ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 // use case : calculate length || magnitude of a single vector
 
-double	vec4_len(t_vec4 vec)
+float	vec4_len(t_vec4 vec)
 {
-	double	len;
+	float	len;
 
 	len = sqrt(vec4_dot_prod(vec, vec));
 	return (len);
