+++ src/vec4/vec4_norm.c
@@ -6,7 +6,7 @@
 /*   By: rshin <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/02 14:47:59 by rshin             #+#    #+#             */
-/*   Updated: 2025/10/02 16:06:50 by rshin            ###   ########.fr       */
+/*   Updated: 2025/10/14 20:11:25 by rshin            ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 t_vec4	vec4_norm(t_vec4 vec)
 {
 	t_vec4	res;
-	double	len;
+	float	len;
 
 	len = vec4_len(vec);
 	if (len == 0)
